# docs/api-collection/codesign_swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code Sign API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{domain}",
      "variables": {
        "protocol": { "default": "https" },
        "domain": { "default": "example.com" }
      }
    }
  ],
  "paths": {
...
    "/api/v1/sign": {
      "post": {
        "summary": "Sign hash",
        "operationId": "signHash",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "license_key": { "type": "string" },
                  "activation_id": { "type": "string" },
                  "algorithm": { "type": "string", "enum": ["sha256"] },
                  "hash": { "type": "string" }
                },
                "required": [
                  "license_key",
                  "activation_id",
                  "algorithm",
                  "hash"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed hash",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signature": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}